Make a generic odML parser for XML, YAML and JSON
